добавил id user в контроллер book и на view '/' и '/books/show' исправил немного fron-end. Добавил поле в класс UserDetailsServiceImp

+++ bookshop/src/main/java/ua/step/bookshop/repositories/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ua.step.bookshop.models.User;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Short> {
     Optional<User> findByLogin(String login);
 
 }
